requirements.txt
Flask==2.2.3
Werkzeug<3.0.0
lxml>=4.6.5,<5.0.0
python-dotenv==1.0.0
WTForms==3.0.1
Flask-WTF==1.1.1
email_validator==2.0.0
gunicorn==20.1.0
pytest==7.3.1
requests==2.31.0
Flask-Session==0.4.1
